Add files via upload
The python script "Pipeline.py" contains the code to sort the fit files in certain directories, while "analysis.py" contains the code to reduce the images, register them, extract the photometry of the cluster from the registered images, make color-magnitude diagrams and determine the age and distance of the cluster. 
The jupyter notebooks are the supporting notebooks that contains more detail codes.